+++ CHANGELOG.md
@@ -36,6 +36,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Bug preventing any type of playback due to spotify API changes.
 - Bug preventing retrieval of new song metadata from spotify.
+## [Unreleased]
+
+### Changed
+
+### Added
 
 ## [1.3.0]